refactor: separate into individual crates
Fixes #7

Separate the `car-thing` binary and library into individual crates.

* Modify `Cargo.toml` to add `car-thing-bin` and `car-thing-lib` to the `members` array.
* Rename `crates/car-thing/Cargo.toml` to `crates/car-thing-bin/Cargo.toml` and update the package name to `car-thing-bin`. Add dependency for `car-thing-lib`.
* Rename `crates/car-thing/src/main.rs` to `crates/car-thing-lib/src/lib.rs` and remove the binary implementation.
* Add `crates/car-thing-bin/src/main.rs` to implement the `car-thing-bin` binary using `clap` and `car-thing-lib`.
* Add `crates/car-thing-lib/Cargo.toml` to define the `car-thing-lib` package and add dependency for `rusb`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/tinted-software/car-thing-tools/issues/7?shareId=XXXX-XXXX-XXXX-XXXX).
