+++ crates/car-thing-bin/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "car-thing"
+name = "car-thing-bin"
 version = "0.1.0"
 edition = "2021"
 description = "A tool to interact with Car Thing devices"
 
 [dependencies]
 clap = { version = "4.5.11", features = ["derive"] }
-rusb = "0.9.4"
+car-thing-lib = { path = "../car-thing-lib" }
 
 [lints]
 workspace = true
