+++ crates/car-thing-lib/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "car-thing-lib"
+version = "0.1.0"
+edition = "2021"
+description = "A library to interact with Car Thing devices"
+
+[dependencies]
+rusb = "0.9.4"
+
+[lints]
+workspace = true
